pyproject.toml
[project]
name = "rag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "chromadb>=1.3.0",
    "faiss-cpu>=1.12.0",
    "google-generativeai>=0.8.5",
    "ipykernel>=7.1.0",
    "langchain>=1.0.2",
    "langchain-community>=0.4.1",
    "langchain-core>=1.0.1",
    "pymupdf>=1.26.5",
    "pypdf>=6.1.3",
    "python-dotenv>=1.2.1",
    "sentence-transformers>=5.1.2",
]
